Move bracket component to its own folder.
